002- Directory refactor
organizing directory
