+++ src/App.js
@@ -2,7 +2,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap";
 import React from "react";
 
-import Navbar from "./components/Navbar/Navbar";
+import Navbar from "./components/common/navbar/Navbar";
 
 //Routing
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
