+++ src/components/common/navbar/Navbar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { NavLink, Link } from "react-router-dom";
-import "../../App.scss";
+import "./App.scss";
 
 const Navbar = () => {
   return (
